requirements.txt
Django==2.1.1
mysqlclient==1.3.13
pytz==2018.5
SQLAlchemy==1.2.11
django-multiselectfield==0.1.8